remove warning

+++ trunk/base/grid.c
@@ -160,7 +160,7 @@ void Grid::init_from_atom(const t_atom &x) {
 	} else if (x.a_type==A_FLOAT) {
 		init(new Dim(),int32_e);
 		CHECK_ALIGN2(this->data,nt);
-		((int32 *)*this)[0] = x.a_float;
+		((int32 *)*this)[0] = (int32)x.a_float;
 	} else {
 		rb_funcall(
 		EVAL("proc{|x| raise \"can't convert to grid: #{x.inspect}\"}"),
